## The scene
I've done two pairs of benchmark - before and after of resource usage and a breakdown of running times.
The benchmark is pretty simple but very effective - one browser was just receiving, the other was sending events.
The receiving side was profiled for 20 seconds each time under heavy usage of the sender (a lot of events generated per second).

## Running times profiling
|before|after|
|---|---|
![image](https://github.com/user-attachments/assets/389afb0e-e86e-4c9a-8f23-8fc7bd780114)|![image](https://github.com/user-attachments/assets/12cc13a1-da0b-4607-9fd8-f0b30db05406)

Looking at the `before` the reconciliation took `62.2%` (4486ms) of the total microtask time (11560ms).
The main issue I saw was that `loadFiles` took `99%` of the running time of the reconciliation (4452ms)
The profiling also highlighted the Excalidraw native `updateScene` takes half of the running time which is also a good candite for optimization. The only thing I can think of is calling it less but I suspect that will be hard to do - throttling the incoming events might be an improvement.

## The result
After a simple optimization of `loadFiles` I managed to take down it's running time down to 46ms - **97x!!!!** improvement; the total microtask time to 6803ms (41.2% improvement :arrow_upper_right:) and the reconciliation time **31x times!!!!** (144ms).



|---|before|after|improvement
|---|---|---|---
microtask|11560.4|6803.9|1.7x
reconcileElementsAndLoadFiles|4486.6|144.6|31x
loadFiles|4452.3|46.0|97x


---

## Resource profiling
|before|after|
|---|---|
![image](https://github.com/user-attachments/assets/0d41ad46-ff9f-4b71-95c7-5b6270951e83)|![image](https://github.com/user-attachments/assets/0a0a8b97-a4ce-4d8b-addd-ba1cc121d04d)